# docs/openapi.json
{
  "components": {
    "schemas": {
      "ArrivalInfo": {
        "description": "One route's next-arrival estimate at a stop (issue #48, Seoul TOPIS).",
        "properties": {
          "arrival_message_1": {
            "title": "Arrival Message 1",
            "type": "string"
          },
          "arrival_message_2": {
            "title": "Arrival Message 2",
            "type": "string"
          },
          "congestion_1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Congestion 1"
          },
          "congestion_2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Congestion 2"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "route_id": {
            "title": "Route Id",
            "type": "string"
          },
          "route_name": {
            "title": "Route Name",
            "type": "string"
          }
        },
        "required": [
          "route_name",
          "route_id",
          "direction",
          "arrival_message_1",
          "arrival_message_2",
          "congestion_1",
          "congestion_2"
        ],
        "title": "ArrivalInfo",
        "type": "object"
      },
      "ArrivalsResponse": {
        "description": "Real-time arrivals for every route serving one stop.\n\nIndependent of the wait-population estimate: ``available=False`` means\nTOPIS couldn't be reached (timeout/error).",
        "properties": {
          "arrivals": {
            "items": {
              "$ref": "#/components/schemas/ArrivalInfo"
            },
            "title": "Arrivals",
            "type": "array"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "stop_id": {
            "title": "Stop Id",
            "type": "integer"
          }
        },
        "required": [
          "stop_id",
          "name",
          "available",
          "message",
          "arrivals"
        ],
        "title": "ArrivalsResponse",
        "type": "object"
      },
      "CongestionResponse": {
        "description": "Estimated wait and congestion grade for one stop at one hour.\n\ngrade_basis says what the 여유/보통/혼잡 label is relative to --\n\"seoul_percentile\" means Seoul-wide 70/90th percentiles of W, not the\nstop's own capacity (estimator/congestion.py).",
        "properties": {
          "estimated_wait": {
            "title": "Estimated Wait",
            "type": "number"
          },
          "grade": {
            "title": "Grade",
            "type": "string"
          },
          "grade_basis": {
            "title": "Grade Basis",
            "type": "string"
          },
          "hour": {
            "title": "Hour",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "stop_id": {
            "title": "Stop Id",
            "type": "integer"
          }
        },
        "required": [
          "stop_id",
          "name",
          "hour",
          "estimated_wait",
          "grade",
          "grade_basis"
        ],
        "title": "CongestionResponse",
        "type": "object"
      },
      "CorridorResponse": {
        "properties": {
          "hour": {
            "title": "Hour",
            "type": "integer"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/CorridorStopSnapshot"
            },
            "title": "Stops",
            "type": "array"
          }
        },
        "required": [
          "hour",
          "stops"
        ],
        "title": "CorridorResponse",
        "type": "object"
      },
      "CorridorStopSnapshot": {
        "properties": {
          "estimated_wait": {
            "title": "Estimated Wait",
            "type": "number"
          },
          "grade": {
            "title": "Grade",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "stop_id": {
            "title": "Stop Id",
            "type": "integer"
          }
        },
        "required": [
          "stop_id",
          "name",
          "estimated_wait",
          "grade"
        ],
        "title": "CorridorStopSnapshot",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Stop": {
        "properties": {
          "ars_number": {
            "title": "Ars Number",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "stop_id": {
            "title": "Stop Id",
            "type": "integer"
          },
          "stop_type": {
            "title": "Stop Type",
            "type": "string"
          }
        },
        "required": [
          "stop_id",
          "name",
          "ars_number",
          "lat",
          "lon",
          "stop_type"
        ],
        "title": "Stop",
        "type": "object"
      },
      "StopContextResponse": {
        "description": "Weather + day-type context for one stop on one date (issue #47).\n\nprecipitation_type is a normalized display label (맑음/비/눈 등). is_forecast=True\nmeans the fields came from a short-range forecast rather than an observed\nreading; precipitation/snowfall are 0.0 on forecast rows. wind_speed is\nnull on the rare historical date where ASOS itself has no 평균풍속\nreading (issue #137) -- never fabricated, since unlike precipitation/\nsnowfall there's no \"0 means it didn't happen\" convention for wind.",
        "properties": {
          "congestion_note": {
            "title": "Congestion Note",
            "type": "string"
          },
          "date": {
            "title": "Date",
            "type": "integer"
          },
          "day_type": {
            "title": "Day Type",
            "type": "string"
          },
          "humidity": {
            "title": "Humidity",
            "type": "number"
          },
          "is_forecast": {
            "title": "Is Forecast",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "precipitation": {
            "title": "Precipitation",
            "type": "number"
          },
          "precipitation_type": {
            "title": "Precipitation Type",
            "type": "string"
          },
          "snowfall": {
            "title": "Snowfall",
            "type": "number"
          },
          "stop_id": {
            "title": "Stop Id",
            "type": "integer"
          },
          "temperature": {
            "title": "Temperature",
            "type": "number"
          },
          "wind_speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed"
          }
        },
        "required": [
          "stop_id",
          "name",
          "date",
          "day_type",
          "temperature",
          "precipitation",
          "humidity",
          "snowfall",
          "wind_speed",
          "precipitation_type",
          "is_forecast",
          "congestion_note"
        ],
        "title": "StopContextResponse",
        "type": "object"
      },
      "StopsResponse": {
        "properties": {
          "stops": {
            "items": {
              "$ref": "#/components/schemas/Stop"
            },
            "title": "Stops",
            "type": "array"
          }
        },
        "required": [
          "stops"
        ],
        "title": "StopsResponse",
        "type": "object"
      },
      "TimelinePoint": {
        "properties": {
          "estimated_wait": {
            "title": "Estimated Wait",
            "type": "number"
          },
          "grade": {
            "title": "Grade",
            "type": "string"
          },
          "hour": {
            "title": "Hour",
            "type": "integer"
          }
        },
        "required": [
          "hour",
          "estimated_wait",
          "grade"
        ],
        "title": "TimelinePoint",
        "type": "object"
      },
      "TimelineResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "stop_id": {
            "title": "Stop Id",
            "type": "integer"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelinePoint"
            },
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "stop_id",
          "name",
          "timeline"
        ],
        "title": "TimelineResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Station Cast API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/corridor": {
      "get": {
        "description": "Every stop's estimated wait and congestion grade at one hour (default: current hour).\n\ndata.wait's [표준버스정류장ID, 시간대] index (api/data.py) makes the\nper-hour cross-section an O(1)-ish index lookup instead of a full-table\nscan. Grading needs no per-stop lookup at all now that thresholds are\nSeoul-wide -- they're read once, before the loop.",
        "operationId": "get_corridor_api_corridor_get",
        "parameters": [
          {
            "in": "query",
            "name": "hour",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 23,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hour"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorridorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Corridor"
      }
    },
    "/api/stops": {
      "get": {
        "description": "List stops in the corridor, optionally viewport-filtered and paginated.\n\nNo bbox/limit returns every stop, same as before -- both are opt-in so\nSTATIONCAST_SCOPE=seoul's ~11,000 stops don't become a required query\nparam for demo-scope callers that don't need them yet.",
        "operationId": "list_stops_api_stops_get",
        "parameters": [
          {
            "description": "minlon,minlat,maxlon,maxlat",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "minlon,minlat,maxlon,maxlat",
              "title": "Bbox"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Stops"
      }
    },
    "/api/stops/{stop_id}/arrivals": {
      "get": {
        "description": "Real-time next-arrival info for every route serving one stop (issue #48).\nBacked by Seoul TOPIS (ws.bus.go.kr) -- a live display feature.",
        "operationId": "get_arrivals_api_stops__stop_id__arrivals_get",
        "parameters": [
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "title": "Stop Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrivalsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Arrivals"
      }
    },
    "/api/stops/{stop_id}/congestion": {
      "get": {
        "description": "Estimated wait and congestion grade for one stop at one hour (default: current hour).",
        "operationId": "get_congestion_api_stops__stop_id__congestion_get",
        "parameters": [
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "title": "Stop Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "hour",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 23,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hour"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CongestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Congestion"
      }
    },
    "/api/stops/{stop_id}/context": {
      "get": {
        "description": "Weather + day-type context for one stop on one date (default: today).\n\nHistorical dates (within weather_daily.parquet's collection window) use\nobserved ASOS data. Dates outside that window -- typically \"오늘\" once the\nwindow rolls past its end date -- fall back to KMA's short-range forecast\n(ingest/weather_forecast.py), flagged is_forecast=True.",
        "operationId": "get_context_api_stops__stop_id__context_get",
        "parameters": [
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "title": "Stop Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Context"
      }
    },
    "/api/stops/{stop_id}/timeline": {
      "get": {
        "description": "Full 24-hour estimated-wait curve for one stop, each hour graded.",
        "operationId": "get_timeline_api_stops__stop_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "stop_id",
            "required": true,
            "schema": {
              "title": "Stop Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Timeline"
      }
    },
    "/health": {
      "get": {
        "description": "Health check endpoint.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}